Check for permissions

+++ server/methods/resetAvatar.js
@@ -1,12 +1,13 @@
 import { Meteor } from 'meteor/meteor';
 import { DDPRateLimiter } from 'meteor/ddp-rate-limiter';
 import { FileUpload } from '../../app/file-upload';
-import { Users } from '../../app/models';
+import { Users } from '../../app/models/server';
 import { settings } from '../../app/settings';
 import { Notifications } from '../../app/notifications';
+import { hasPermission } from '../../app/authorization/server';
 
 Meteor.methods({
-	resetAvatar() {
+	resetAvatar({ userId }) {
 		if (!Meteor.userId()) {
 			throw new Meteor.Error('error-invalid-user', 'Invalid user', {
 				method: 'resetAvatar',
@@ -19,30 +20,30 @@ Meteor.methods({
 			});
 		}
 
-		const user = Meteor.user();
+		let user;
+
+		if (userId) {
+			if (!hasPermission(userId, 'edit-other-user-avatar')) {
+				throw new Meteor.Error('error-unauthorized', 'Unauthorized', {
+					method: 'resetAvatar',
+				});
+			}
+
+			user = Users.findOneById(userId, { fields: { _id: 1, username: 1 } });
+		} else {
+			user = Meteor.user();
+		}
+
+		if (user == null) {
+			throw new Meteor.Error('error-invalid-desired-user', 'Invalid desired user', {
+				method: 'resetAvatar',
+			});
+		}
+
 		FileUpload.getStore('Avatars').deleteByName(user.username);
 		Users.unsetAvatarOrigin(user._id);
 		Notifications.notifyLogged('updateAvatar', {
 			username: user.username,
-		});
-	},
-	resetAvatarAdmin(args) {
-		if (!args) {
-			throw new Meteor.Error('error-invalid-user', 'Invalid user', {
-				method: 'resetAvatarAdmin',
-			});
-		}
-
-		if (!settings.get('Accounts_AllowUserAvatarChange')) {
-			throw new Meteor.Error('error-not-allowed', 'Not allowed', {
-				method: 'resetAvatarAdmin',
-			});
-		}
-
-		FileUpload.getStore('Avatars').deleteByName(args.username);
-		Users.unsetAvatarOrigin(args._id);
-		Notifications.notifyLogged('updateAvatar', {
-			username: args.username,
 		});
 	},
 });
